ci: run ci on java 8
Just found out that the Gradle dependency manager doesn't respect the actual version of a library, rather the `org.gradle.jvm.version` attribute. By default this is set to the Gradle JVM version and since the new JVM 8 targeted plugin was still being built and published on Java 11 Gradle would still think it's on Java 11 making the change useless.