docs: note that the 'scopes' key is a list of (bash) regexes
As noted in #442, some users may want to allow e.g. uppercase letters
or numbers in their commit scope. This is not formally supported by the
conventional commit standard but achievable by adding the relevant
regex to the `scopes` configuration key. Unfortunately, it was not
clear from the documentation that this key accepts (bash-style)
regexes, making it hard for users to discover this functionality and
risky to use in case of future changes to the implementation.

This commit explicitly notes that the `scopes` configuration key
accepts bash-style regexes. It also changes the listed default for this
key to be `listOf("[a-z \-]")`. While this last part may not be
completely true (it should actually be `emptyList()`), this change does
make the default clear to skimming readers.